Change egd_get_cap and add egd_get_numch API functions
egd_get_cap is modified to avoid passing structure to the interface. This
leads to a safer use of the library (not depending of compiler flags) as
well as a better future extensibility of the library. Since future recording
device may sample new sensor type, the API call egd_get_numch has been
created to query that information.