gtec: Add real support for multiple systems (major rework)
Now deviceid option support the concatanation of several device ids with '+'
symbols. The first device is configured as master and all the other as
slaves. This combination is exposed to the user of the library as a unique
entity that provides 16+1 channels multiplied by the number of devices. All
channels are obviously synchronized.