Separate plugin code from core library implementation
This commit do not export core library implementation into the definition of
the module specific structure. Instead, the structure eegdev exports only
one of its fields (devmodule) which contains the core interface functions
that the plugin can call. The device specific structure is constructed from
it (next to it in the structure layout). This separation avoid to break the
plugin ABI/API whenever a change is nade in the core library.