biosemi: fix race conditions in tests
Fix two race conditions in the synchronous transfer and in transfer queues
destruction of the fake libusb library. In addition to these fixes, the
synchronous transfer wrapper has slightly been rewritten to avoid the use of
stack allocated objects for synchronization: this is done solely to avoid
the false positives with helgrind (which have masked until now the previous
race conditions).