biosemi: fix possible hangups of egd_close
When stopping the handshake, URB can report of URB failure because of
timeout. This patch makes sure that whenever a URB is not resubmitted, this
will be indicated by reducing the number of running transfers
(a2dev->num_running). This should also fix the hang when closing a device
that has failed during acquisition.