gtec: Allow to dlopen gusbampapi library
Adds a configure option to control how the gtec backend access libgusbampapi
library. It can be either by linking at compilation time as it is now, or it
can try to dlopen it at runtime. This later option allows among other to
build the gtec plugin in the absence of the libgusbampapi and allows to use
it on other systems that have the library (ie. the build system does not
need libgusbampi).