biosemi: try to realign to first sync at startup
When communication with device start, if there has been a previous
acquisition, it is possible that the frame will not start at the beginning
of the buffer. In order to handle this, we try to find the first synchro
code in the data transferred and to realign to this one.