datafile: Add loop option
Provides a new option "loop" to datafile plugin. If set to a
non-negative integer, when the end of file is reached, it will be
rewinded instead of reporting an error as many time as the integer
passed as value. If set to "yes" or "true", the file is rewinded
infinitively. Never if set to "no" or "false".

If loop option is not set, it is assumed to be "no".

Closes: MMLAB-91
Change-Id: Iaf4ce9b07e476377aab842114f6b06b927442be2