mmpack: disable lto as a workaround
Because meson combines the compilation of object file meant for static
linking and some meant for shared library (sensortype.c), there are
issue of relocation when activation LTO on meson. The real culprit is
how the tests are implemented on meson, but this would requires more
refactoring.

Hence as a workaround, until a proper fix can be issued, the link time
optimization (LTO) is disabled when build the mmpack packages.

Change-Id: Icddaafaf0bfd197fc4cdaed0192b8a457b1bd809